done-week3-4

+++ app/Services/AuthService.php
@@ -13,34 +13,42 @@ class AuthService extends BaseService
         parent::__construct($customerRepository);
         $this->customerRepository=$customerRepository;
     }
-    public function register($data){
-        
-        $customer = $this->store($data);
-
-        return $customer;
-        
-
-    }
-    public function login($data){
+    public function login($data,$isWeb=null){
         // check if user phone is already registered
         $customer = $this->customerRepository->checkByPhone($data['phone']);
 
         if (!$customer || !Hash::check($data['password'],$customer->password)){
-            return ['errorCustomer'=>true];
+            
+                return ['errorCustomer'=>true];
+            
+        }
+        if(!$isWeb){
+            $token = $this->customerRepository->createToken($customer);
+            return ['token'=>$token,'customer'=>$customer];
+        }else{
+            Auth::guard('web-customer')->login($customer);
         }
 
-        $token = $this->customerRepository->createToken($customer);
-
-        return ['token'=>$token,'customer'=>$customer];
     }
 
-    public function profile(){
-        $customer = Auth::guard('api-customer')->user();
+    public function profile($isWeb=null){
+        if($isWeb){
+            $customer = Auth::guard('web-customer')->user();
+
+        }else{
+
+            $customer = Auth::guard('api-customer')->user();
+        }
         return $customer;
     }
-    public function logout(){
-        $customer = $this->profile();
-        $this->customerRepository->deleteToken($customer);
+    public function logout($isWeb=null){
+        if($isWeb){
+            $customer = $this->profile(true);
+            Auth::guard('web-customer')->logout();
+        }else{
+            $customer = $this->profile();
+            $this->customerRepository->deleteToken($customer);
+        }
     }
 
 }
